requirements.txt
# Core Flask
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Email validation
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# Security
Werkzeug==3.0.1

# Database
# Note: Version 2.0.23 is incompatible with Python 3.13. 
# Using >=2.0.30 to ensure stability.
SQLAlchemy>=2.0.30

# Production server
gunicorn==21.2.0

# Existing dependencies (from your project)
anthropic>=0.3.0
google-api-python-client>=2.100.0
praw>=7.7.0
requests>=2.31.0
pytrends>=4.9.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
tenacity>=8.2.3
playwright>=1.40.0
pytest-playwright>=0.4.0
